create CFGAnalyzer and set up class hierarchy. Remove analyze functionality from Utterance. Remove chat from Analyzer init
