+++ src/cltl/triple_extraction/api.py
@@ -1,28 +1,24 @@
-import json
 import logging
-import os
 import sys
 from collections import Counter
 from datetime import datetime
 from random import getrandbits
 from typing import List, Optional
 
-from nltk import CFG, RecursiveDescentParser, edit_distance
+from nltk import edit_distance
 from nltk import pos_tag
 
 from cltl.combot.backend.api.discrete import UtteranceType
 from cltl.combot.backend.utils.casefolding import casefold_text
-from cltl.triple_extraction.analyzer import Analyzer
 from cltl.triple_extraction.data.base_cases import friends
-from cltl.triple_extraction.ner import NER
-from cltl.triple_extraction.pos import POS
+from cltl.triple_extraction.nlp.parser import Parser
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
 
 handler = logging.StreamHandler(sys.stdout)
 handler.setLevel(logging.INFO)
-formatter = logging.Formatter('\r%(asctime)s - %(levelname)8s - %(name)40s - %(message)s')
+formatter = logging.Formatter('\r%(asctime)s - %(levelname)8s - %(name)60s - %(message)s')
 handler.setFormatter(formatter)
 logger.addHandler(handler)
 
@@ -206,12 +202,12 @@ class Utterance(object):
         self._hypothesis = self._choose_hypothesis(hypotheses)
         self._tokens = self._clean(self._tokenize(self.transcript))
 
-        # TODO: Optimize: takes 2.6 seconds now! Should be < 1 second!?
-        self._parser = Parser(self)
-        # TODO analyze sets triple, perspective and type, but currently is not called on constructor
         self._type = None
         self._triple = None
         self._perspective = None
+
+        # TODO: Optimize: takes 2.6 seconds now! Should be < 1 second!?
+        self._parser = Parser(self)
 
     @property
     def chat(self):
@@ -279,9 +275,6 @@ class Utterance(object):
         """
         return self._turn
 
-    def set_turn(self, value):
-        self._turn = value
-
     @property
     def triple(self):
         # type: () -> dict
@@ -338,28 +331,12 @@ class Utterance(object):
         """
         return self._parser
 
-    def analyze(self):
-        """
-        Determines the type of utterance, extracts the RDF triple and perspective attaching them to the last utterance
-        Returns
-        -------
+    def set_turn(self, value):
+        self._turn = value
 
-        """
-        # Analyze utterance
-        analyzer = Analyzer.analyze(self._chat)
-
-        if not analyzer:
-            return "I cannot parse your input"
-
-        for el in ["subject", "predicate", "object"]:
-            Analyzer.LOG.info(
-                "RDF {:>10}: {}".format(el, json.dumps(analyzer.triple[el], sort_keys=True, separators=(', ', ': '))))
-
-        # Set type, triple and perspective
-        self._type = analyzer.utterance_type
-        self.set_triple(analyzer.triple)
-        if analyzer.utterance_type == UtteranceType.STATEMENT:
-            self.set_perspective(analyzer.perspective)
+    def set_type(self, utterance_type):
+        # type: (UtteranceType) -> ()
+        self._type = utterance_type
 
     def set_triple(self, triple):
         # type: (dict) -> ()
@@ -526,122 +503,3 @@ class Utterance(object):
     def __repr__(self):
         author = self.chat.speaker
         return '{:>10s}: "{}"'.format(author, self.transcript)
-
-
-class Parser(object):
-    POS_TAGGER = None  # Type: POS
-    NER_TAGGER = None
-    GRAMMAR = None
-    CFG_GRAMMAR_FILE = os.path.join(os.path.dirname(__file__), 'data', 'cfg.txt')
-
-    def __init__(self, utterance):
-        self._log = logger.getChild(self.__class__.__name__)
-
-        if not Parser.POS_TAGGER:
-            Parser.POS_TAGGER = POS()
-            logger.info("Started POS tagger")
-
-        if not Parser.NER_TAGGER:
-            Parser.NER_TAGGER = NER()
-            logger.info("Started NER tagger")
-
-        with open(Parser.CFG_GRAMMAR_FILE) as cfg_file:
-            if not Parser.GRAMMAR:
-                Parser.GRAMMAR = cfg_file.read()
-                logger.info("Loaded grammar")
-            self._cfg = Parser.GRAMMAR
-
-        self._forest, self._constituents = self._parse(utterance)
-
-    @property
-    def forest(self):
-        return self._forest
-
-    @property
-    def constituents(self):
-        return self._constituents
-
-    def _parse(self, utterance):
-        """
-        :param utterance: an Utterance object, typically last one in the Chat
-        :return: parsed syntax tree and a dictionary of syntactic realizations
-        """
-        self._log.debug("Start parsing")
-        tokenized_sentence = utterance.tokens
-        pos = self.POS_TAGGER.tag(tokenized_sentence)  # standford
-        alternative_pos = pos_tag(tokenized_sentence)  # nltk
-
-        self._log.debug(pos)
-        self._log.debug(alternative_pos)
-
-        if pos != alternative_pos:
-            self._log.debug('DIFFERENT POS tag: %s != %s' % (pos, alternative_pos))
-
-        # fixing issues with POS tagger (Does and like)
-        ind = 0
-        for w in tokenized_sentence:
-            if w == 'like':
-                pos[ind] = (w, 'VB')
-            ind += 1
-
-        if pos and pos[0][0] == 'Does':
-            pos[0] = ('Does', 'VBD')
-
-        # the POS tagger returns one tag with a $ sign (POS$) and this needs to be fixed for the CFG parsing
-        ind = 0
-        for word, tag in pos:
-            if '?' in word:
-                word = word[:-1]
-            if tag.endswith('$'):
-                new_rule = tag[:-1] + 'POS -> \'' + word + '\'\n'
-                pos[ind] = (pos[ind][0], 'PRPPOS')
-            else:
-                # CFG grammar is created dynamically, with the terminals added each time from the specific utterance
-                new_rule = tag + ' -> \'' + word + '\'\n'
-            if new_rule not in self._cfg:
-                self._cfg += new_rule
-            ind += 1
-
-        try:
-            cfg_parser = CFG.fromstring(self._cfg)
-            rd = RecursiveDescentParser(cfg_parser)
-
-            last_token = tokenized_sentence[len(tokenized_sentence) - 1]
-
-            if '?' in last_token:
-                tokenized_sentence[len(tokenized_sentence) - 1] = last_token[:-1]
-
-            parsed = rd.parse(tokenized_sentence)
-
-            s_r = {}  # syntactic_realizations are the topmost branches, usually VP/NP
-            index = 0
-
-            forest = [tree for tree in parsed]
-
-            if len(forest):
-                if (len(forest)) > 1:
-                    self._log.debug('* Ambiguity in grammar *')
-                for tree in forest[0]:  # alternative trees? f
-                    for branch in tree:
-                        s_r[index] = {'label': branch.label(), 'structure': branch}
-                        raw = ''
-                        for node in branch:
-                            for leaf in node.leaves():
-                                raw += leaf + '-'
-
-                        s_r[index]['raw'] = raw[:-1]
-                        index += 1
-            else:
-                self._log.debug("no forest")
-
-            for el in s_r:
-                if type(s_r[el]['raw']) == list:
-                    string = ''
-                    for e in s_r[el]['raw']:
-                        string += e + ' '
-                    s_r[el]['raw'] = string
-                s_r[el]['raw'] = s_r[el]['raw'].strip()
-
-            return forest, s_r
-        except:
-            return [], {}
